producer/requirements.txt
requests
protobuf
gtfs-realtime-bindings
kafka-python